Bug KGF-155 all the functions were moved to functions.py, imported to necessary places. It helped to reduce the duplicate of code. All the lists of players were sorted as it was requested by the task.
