+++ source/webapp/templates/partial/player_list.html
@@ -13,7 +13,7 @@
     </tr>
     </thead>
     <tbody>
-    {% for el in sorted_players %}
+    {% for el in players %}
         <tr>
             <td>{{ forloop.counter }}</td>
             <td><a style="text-decoration: underline dotted ghostwhite; color: black;"
