+++ source/webapp/templates/player/player_search.html
@@ -40,11 +40,12 @@
     <br/>
     <div class="container">
         {% include 'partial/player_list.html' %}
-    </div>
+
     <br/>
     {% if is_paginated %}
         {% include 'partial/pagination.html' %}
     {% endif %}
+    </div>
     <br/>
     </div>
