+++ source/webapp/views/club_view.py
@@ -1,42 +1,12 @@
 import re
 from urllib.parse import urlencode
-
 from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from django.urls import reverse
-
 from webapp.forms import ClubSearch, ClubForm
-from webapp.models import Club, Country, PlayerInTournament
+from webapp.models import Club
 from django.views.generic import ListView, TemplateView, UpdateView
-
-
-def average_go_level():
-    # Здесь нужно будет привязать фильтр через страну клуба, чтобы выводил только по Кыргызстану
-    clubs = Club.objects.all()
-    club_list = []
-    for club in clubs:
-        new_dict = dict()
-        num_list = []
-        num_players = club.players.count()
-        for player in club.players.all():
-            tournament = player.tournaments.all().order_by('-date')[:1]
-            data = player.playerintournament_set.get(tournament_id=tournament)
-            if data.GoLevel:
-                p = re.compile('(\d*)')
-                m = p.findall(data.GoLevel)
-                for i in m:
-                    if i != "":
-                        num_list.append(int(i))
-        total_go_level = sum(num_list)
-        if num_players > 0:
-            average_num = total_go_level // num_players
-            new_dict['club'] = club.pk
-            new_dict['average'] = average_num
-        else:
-            new_dict['club'] = club.pk
-            new_dict['average'] = 0
-        club_list.append(new_dict)
-    return club_list
+from webapp.views.functions import get_list_of_filtered_players, get_rank, average_go_level, sorted_list_of_players
 
 
 class ClubsListView(ListView):
@@ -92,18 +62,10 @@ class ClubView(TemplateView):
         pk = kwargs.get("pk")
         club = get_object_or_404(Club, pk=pk)
         players = club.players.all()
+        data = get_rank(players)
+        sorted_players = get_list_of_filtered_players(data, sorted_list_of_players)
         kwargs["club"] = club
-        kwargs['players'] = players
-        new_list = []
-        for player in players:
-            new_dict = dict()
-            tournament = player.tournaments.all().order_by('-date')[:1]
-            data = player.playerintournament_set.get(tournament_id=tournament)
-            new_dict['player'] = player.pk
-            new_dict['GoLevel'] = data.GoLevel
-            new_list.append(new_dict)
-        kwargs['rating'] = new_list
-        kwargs['num_participants'] = club.players.count
+        kwargs['players'] = sorted_players
         club_list = average_go_level()
         for element in club_list:
             if club.pk == element['club']:
